requirements.txt
Flask
gunicorn
psycopg2-binary
pandas 
openpyxl 
qrcode 
pillow
dotenv